refactor: extract transcribe and vad helpers into shared lib
